Code unification; add Summon HomePage facet support.
