+++ config/vufind/Summon.ini
@@ -225,6 +225,20 @@ orFacets = *
 ; below, no facets will be translated.
 translated_facets[] = ContentType
 
+; These facets will be displayed on the Home Page.  If this section is omitted,
+; the [Advanced_Facets] section will be used instead.
+[HomePage_Facets]
+Language = "Language"
+ContentType = "Format"
+
+; These settings affect the way the [HomePage] facets are displayed
+; NOTE: To make changes take effect immediately, you may need to clear VuFind's
+; cache after changing this section.
+[HomePage_Facet_Settings]
+; how many values should we load for each facet?  depending on the column layout
+; of the homepage facet lists, we may not display all loaded values for every facet
+facet_limit = 100
+
 ; This section shows which search types will display in the basic search box at
 ; the top of Summon pages.  The name of each setting below corresponds with an
 ; index defined in the Summon API.  The value of each setting is the text to
