+++ module/VuFind/src/VuFind/Search/Base/FacetCacheFactory.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Solr FacetCache Factory.
+ * Abstract FacetCache Factory.
  *
  * PHP version 7
  *
@@ -20,7 +20,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  *
  * @category VuFind
- * @package  Search_Solr
+ * @package  Search_Base
  * @author   Demian Katz <demian.katz@villanova.edu>
  * @license  http://opensource.org/licenses/gpl-2.0.php GNU General Public License
  * @link     https://vufind.org/wiki/development Wiki
@@ -28,37 +28,51 @@
 namespace VuFind\Search\Solr;
 
 use Interop\Container\ContainerInterface;
+use Zend\ServiceManager\Factory\FactoryInterface;
 
 /**
- * Solr FacetCache Factory.
+ * Abstract FacetCache Factory.
  *
  * @category VuFind
- * @package  Search_Solr
+ * @package  Search_Base
  * @author   Demian Katz <demian.katz@villanova.edu>
  * @license  http://opensource.org/licenses/gpl-2.0.php GNU General Public License
  * @link     https://vufind.org/wiki/development Wiki
  */
-class FacetCacheFactory extends \VuFind\Search\Base\FacetCacheFactory
+abstract class FacetCacheFactory implements FactoryInterface
 {
     /**
-     * Create a results object with hidden filters pre-populated.
+     * Create a results object with appropriate pre-populated values.
      *
      * @param ContainerInterface $container Service manager
      *
      * @return \VuFind\Search\Base\Results
      */
-    protected function getResults(ContainerInterface $container)
-    {
-        $filters = $container->get('VuFind\Search\SearchTabsHelper')
-            ->getHiddenFilters('Solr');
-        $results = $container->get('VuFind\Search\Results\PluginManager')
-            ->get('Solr');
-        $params = $results->getParams();
-        foreach ($filters as $key => $subFilters) {
-            foreach ($subFilters as $filter) {
-                $params->addHiddenFilter("$key:$filter");
-            }
+    abstract protected function getResults(ContainerInterface $container);
+
+    /**
+     * Create an object
+     *
+     * @param ContainerInterface $container     Service manager
+     * @param string             $requestedName Service being created
+     * @param null|array         $options       Extra options (optional)
+     *
+     * @return object
+     *
+     * @throws ServiceNotFoundException if unable to resolve the service.
+     * @throws ServiceNotCreatedException if an exception is raised when
+     * creating a service.
+     * @throws ContainerException if any other error occurs
+     */
+    public function __invoke(ContainerInterface $container, $requestedName,
+        array $options = null
+    ) {
+        if (!empty($options)) {
+            throw new \Exception('Unexpected options sent to factory.');
         }
-        return $results;
+        $results = $this->getResults($container);
+        $cacheManager = $container->get('VuFind\Cache\Manager');
+        $language = $container->get('Zend\Mvc\I18n\Translator')->getLocale();
+        return new $requestedName($results, $cacheManager, $language);
     }
 }
