+++ module/VuFind/src/VuFind/Search/Summon/FacetCache.php
@@ -39,58 +39,12 @@ namespace VuFind\Search\Summon;
 class FacetCache extends \VuFind\Search\Base\FacetCache
 {
     /**
-     * Perform the actual facet lookup.
+     * Get the namespace to use for caching facets.
      *
-     * @return array
+     * @return string
      */
-    protected function getFacetResults()
+    protected function getCacheNamespace()
     {
-        // Check if we have facet results cached, and build them if we don't.
-        $cache = $this->cacheManager->getCache('object', 'summon-facets');
-        $cacheKey = $this->language;
-        if (!($list = $cache->getItem($cacheKey))) {
-            $limit = $this->config->Advanced_Facet_Settings->facet_limit ?? 100;
-            $params = $this->results->getParams();
-            $facetsToShow = $this->config->Advanced_Facets
-                 ?? ['Language' => 'Language', 'ContentType' => 'Format'];
-            $orFields = isset($this->config->Advanced_Facet_Settings->orFacets)
-                ? array_map(
-                    'trim',
-                    explode(',', $this->config->Advanced_Facet_Settings->orFacets)
-                ) : [];
-            foreach ($facetsToShow as $facet => $label) {
-                $useOr = (($orFields[0] ?? '') == '*')
-                    || in_array($facet, $orFields);
-                $params->addFacet(
-                    $facet . ',or,1,' . $limit, $label, $useOr
-                );
-            }
-
-            // We only care about facet lists, so don't get any results:
-            $params->setLimit(0);
-
-            // force processing for cache
-            $list = $this->results->getFacetList();
-
-            $cache->setItem($cacheKey, $list);
-        }
-
-        return $list;
-    }
-
-    /**
-     * Return facet information. This data may come from the cache.
-     *
-     * @param string $context Context of list to retrieve ('Advanced' or 'HomePage')
-     *
-     * @return array
-     */
-    public function getList($context = 'Advanced')
-    {
-        if (!in_array($context, ['Advanced', 'HomePage'])) {
-            throw new \Exception('Invalid context: ' . $context);
-        }
-        // For now, all contexts are handled the same way.
-        return $this->getFacetResults();
+        return 'summon-facets';
     }
 }
